Refactor and move FirstProcessing experiment and fix class searching
The class search for experiments previously assumed every experiment
must be in the package named after the last word of it's name. That is
not true anymore for many Processing experiments that have been moved to
the repositoryEndpoint package. Update it so that values containing dots
are used as-is, directly representing class names.

Signed-off-by: Diego Araújo <diegoamc90@gmail.com>